+++ tests/DOM/conformance/level1/core/hc_characterdataappenddata.hpp
@@ -82,7 +82,7 @@ class hc_characterdataappenddata : public DOMTestCase<string_type, string_adapto
       Node nameNode;
       CharacterData child;
       String childValue;
-      int childLength;
+	  size_t childLength;
       doc = (Document) baseT::load("hc_staff", true);
       elementList = doc.getElementsByTagName(SA::construct_from_utf8("strong"));
       nameNode = elementList.item(0);
@@ -90,7 +90,7 @@ class hc_characterdataappenddata : public DOMTestCase<string_type, string_adapto
       child.appendData(SA::construct_from_utf8(", Esquire"));
       childValue = child.getData();
       childLength = SA::length(childValue);
-      baseT::assertEquals(24, childLength, __LINE__, __FILE__);
+      baseT::assertEquals(24, static_cast<int>(childLength), __LINE__, __FILE__);
   
    }
